refactoring: created the sub-directory 'http' in kasper-exposition
