+++ kasper-exposition/src/main/java/com/viadeo/kasper/exposition/http/HttpCommandExposerPlugin.java
@@ -4,7 +4,7 @@
 //
 //           Viadeo Framework for effective CQRS/DDD architecture
 // ============================================================================
-package com.viadeo.kasper.exposition;
+package com.viadeo.kasper.exposition.http;
 
 import com.codahale.metrics.MetricRegistry;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -16,6 +16,7 @@ import com.viadeo.kasper.client.platform.domain.descriptor.CommandHandlerDescrip
 import com.viadeo.kasper.client.platform.domain.descriptor.DomainDescriptor;
 import com.viadeo.kasper.cqrs.command.Command;
 import com.viadeo.kasper.cqrs.command.CommandHandler;
+import com.viadeo.kasper.exposition.ExposureDescriptor;
 import com.viadeo.kasper.tools.ObjectMapperProvider;
 
 import java.util.List;
