+++ kasper-exposition/src/main/java/com/viadeo/kasper/exposition/http/HttpExposerError.java
@@ -4,7 +4,7 @@
 //
 //           Viadeo Framework for effective CQRS/DDD architecture
 // ============================================================================
-package com.viadeo.kasper.exposition;
+package com.viadeo.kasper.exposition.http;
 
 public class HttpExposerError extends Error {
