+++ kasper-exposition/src/main/java/com/viadeo/kasper/exposition/http/HttpExposerPlugin.java
@@ -4,7 +4,7 @@
 //
 //           Viadeo Framework for effective CQRS/DDD architecture
 // ============================================================================
-package com.viadeo.kasper.exposition;
+package com.viadeo.kasper.exposition.http;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.viadeo.kasper.client.platform.plugin.Plugin;
