+++ kasper-exposition/src/test/java/com/viadeo/kasper/exposition/BaseHttpExposerTest.java
@@ -15,6 +15,7 @@ import com.viadeo.kasper.client.platform.Platform;
 import com.viadeo.kasper.client.platform.configuration.KasperPlatformConfiguration;
 import com.viadeo.kasper.client.platform.configuration.PlatformConfiguration;
 import com.viadeo.kasper.client.platform.domain.DomainBundle;
+import com.viadeo.kasper.exposition.http.HttpExposerPlugin;
 import com.viadeo.kasper.tools.ObjectMapperProvider;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.servlet.ServletContextHandler;
