+++ kasper-exposition/src/test/java/com/viadeo/kasper/exposition/HttpCommandExposerContextTest.java
@@ -23,6 +23,7 @@ import com.viadeo.kasper.cqrs.query.QueryHandler;
 import com.viadeo.kasper.ddd.Domain;
 import com.viadeo.kasper.ddd.repository.Repository;
 import com.viadeo.kasper.event.EventListener;
+import com.viadeo.kasper.exposition.http.HttpCommandExposerPlugin;
 import org.junit.Test;
 
 import java.util.Locale;
