+++ kasper-exposition/src/test/java/com/viadeo/kasper/exposition/HttpQueryExposerTest.java
@@ -20,6 +20,7 @@ import com.viadeo.kasper.context.impl.DefaultContextBuilder;
 import com.viadeo.kasper.cqrs.query.*;
 import com.viadeo.kasper.cqrs.query.annotation.XKasperQueryHandler;
 import com.viadeo.kasper.cqrs.query.exceptions.KasperQueryException;
+import com.viadeo.kasper.exposition.http.HttpQueryExposerPlugin;
 import com.viadeo.kasper.tools.ObjectMapperProvider;
 import org.junit.Test;
 import org.junit.runner.RunWith;
